+++ src/main/java/com/leets/X/domain/user/dto/response/UserSocialLoginResponse.java
@@ -1,7 +1,7 @@
 package com.leets.X.domain.user.dto.response;
 
 import com.leets.X.domain.user.service.LoginStatus;
-import com.leets.X.global.auth.jwt.JwtResponse;
+import com.leets.X.global.auth.jwt.dto.JwtResponse;
 import lombok.Builder;
 
 @Builder
